go-selinux: move to internal/impl
This moves (git mv) the whole package to under internal, with the
minimal changes:
 - package name;
 - remove the deprecated ReserveLabel;
 - remove the deprecated notice from ContainerLabels (it will stay as
   internal function).

This commit breaks git-bisect; this is done deliberately to show this is
merely the moving files around. The next commit fixes things.

Technically, we can also remove the wrappers and instead make the actual
implementation methods public, but there is no need.
