+++ .github/workflows/validate.yml
@@ -147,7 +147,7 @@ jobs:
       # https://github.com/opencontainers/selinux/issues/225
       - name: "racy test"
         continue-on-error: true
-        run: lima bash -c 'cd /tmp/selinux && go test -timeout 10m -count 100000 ./go-selinux'
+        run: lima bash -c 'cd /tmp/selinux && go test -timeout 10m -count 100000 ./internal/impl'
 
       - name: "Show AVC denials"
         run: lima sudo ausearch -m AVC,USER_AVC || true
