+++ internal/impl/selinux_linux.go
@@ -1,4 +1,4 @@
-package selinux
+package impl
 
 import (
 	"bufio"
